Split out mock utils
